SoSFUS
Checks if the folder called "embeddings" already exists at \mods\xVASynth\resources\app on each launch of SoS.
If "embeddings" doesnt already exist in "\mods\xVASynth\resources\app\embeddings" then it will create the folder there for you. This is most likely temporary and will be obsolete in the next WJ updates. 

Created for Kvitekvist to make the current .bat approach obsolete and FUS more plug and play with Mantella.